+++ ext/FerriteSparseMatrixCSR.jl
@@ -5,16 +5,16 @@ import Ferrite: AbstractSparsityPattern, CSRAssembler, FastSparsityPattern, getn
 import Base: @propagate_inbounds
 
 # Could be generalized if https://github.com/JuliaSparse/SparseArrays.jl/pull/546 is merged
-function Ferrite.start_assemble(K::SparseMatrixCSR{<:Any, T}, f::Vector = T[]; fillzero::Bool = true, maxcelldofs_hint::Int = 0) where {T}
+function Ferrite.start_assemble(K::SparseMatrixCSR{<:Any, T}, f::Vector = T[]; fillzero::Bool = true, maxcelldofs_hint::Int = 0, atomic::Bool = false) where {T}
     fillzero && (Ferrite.fillzero!(K); Ferrite.fillzero!(f))
-    return CSRAssembler(K, f, zeros(Int, maxcelldofs_hint), zeros(Int, maxcelldofs_hint), zeros(Int, maxcelldofs_hint), zeros(Int, maxcelldofs_hint))
+    return CSRAssembler(K, f, zeros(Int, maxcelldofs_hint), zeros(Int, maxcelldofs_hint), zeros(Int, maxcelldofs_hint), zeros(Int, maxcelldofs_hint), atomic)
 end
 
 @propagate_inbounds function Ferrite._assemble_inner!(
         K::SparseMatrixCSR, Ke::AbstractMatrix,
         rowdofs::AbstractVector, sortedrowdofs::AbstractVector, rowpermutation::AbstractVector,
         coldofs::AbstractVector, sortedcoldofs::AbstractVector, colpermutation::AbstractVector,
-        sym::Bool
+        sym::Bool, atomic::Bool
     )
     current_row = 1
     ld = length(coldofs)
@@ -32,7 +32,11 @@ end
                 # Match: add the value (if non-zero) and advance the pointers
                 val = Ke[Kerow, colpermutation[ci]]
                 if !iszero(val)
-                    K.nzval[C] += val
+                    if atomic
+                        Ferrite.atomic_addindex!(K.nzval, val, C)
+                    else
+                        K.nzval[C] += val
+                    end
                 end
                 ci += 1
                 Ci += 1